requirements.txt
# Requires Python 3.10+

# RSS parsing
feedparser==6.0.11

# EPUB creation
ebooklib==0.18

# HTML cleaning (for nicer EPUB articles)
beautifulsoup4==4.12.3
lxml==5.3.0

# Dropbox API
dropbox==12.0.2

# YAML config parsing
pyyaml==6.0.2

# Date/time helpers
python-dateutil==2.9.0.post0

# Logging with colors (optional, but useful)
colorlog==6.8.2

# OAuth 2.0 and security
python-dotenv==1.0.1
cryptography==42.0.5
requests==2.31.0

# Development and testing
pytest==8.1.1
pytest-cov==4.1.0
responses==0.24.1  # For mocking HTTP requests in tests
